Don't lose errno from rpmPackageFilesInstall
There are a lot of ways to lose the value of errno after a failed
rpmPackageFilesInstall.  Typically, stdio will call isatty which then
overwrites errno with ENOTTY.  Save and restore errno between the return
from rpmPackageFilesInstall and the logging of the error.  There may be
more opportunities inside rpmPackageFilesInstall to clobber the real
errno which I haven't investigated yet.

Signed-off-by: Panu Matilainen <pmatilai@redhat.com>