Support rich dependencies in rpmtsCheck
We use rpmdsParseRichDep() to split the rich dependency and then
recursively call unsatisfiedDepend() on the sub-dependencies.

We also need to deal with "negated" dependencies, a requires
with an IF dependency produces conflict like subdependencies.
Those subdependencies are prefixed with a '!' character in the
requires index database. We currently do not support this for
conflict dependencies, this is enforced by the spec file parser.

Fortunately we already read all conflicts/requires from the
index databases, so we can setup hash tables which results in
almost no extra costs.