Fix file signatures issue with sig header size changes
When file signatures are added to a package the signature digests are
replaced. Sometimes the resulting signature header has a different size.
To solve this, gpg reserved space is omitted. This forces the rpm to be
rewritten when file signatures are added.

Changelog:
-no longer effects delsign

(cherry picked from commit 3e730133409ecae86b7f29dd02e31929ffebfac3)