Add future import for print(); Refactor string formatting calls
In case the generator is being used in an environment where
Python 2 is the default still, the appropriate import to ensure
that print() works has been added.

Additionally, the string formatting calls have been refactored
to use string.format() instead of old-style printf-esque calls.
This enhances the readability slightly be making it so that '%'
isn't being used for anything but RPM stuff in strings.