Fix next_brace_sub() to return NULL if braces don't match.
Our implementation of glob() is copied from glibc (2.1 probably) which
has buggy implementation of next_brace_sub(). It is fixed in the newer
version of glibc. Now next_brace_sub() is modified  according to newer
(fixed) version of glic. We didn't reveal this bug sooner because brace
expansion was disabled in rpm but we enabled brace expansion in
commit d14ecfe587efbe80e5534161dbd3a4f7158b4e2b.