Add --legacy to generate pythonegg(X)(M) Provides
Thierry Vignaud brought up on rpm-maint that Mageia
currently uses "pythonegg(X)(M)" (e.g. "pythonegg(3)(rpm)"
for python3 rpm bindings package) in their Python packages
to pull in Python dependencies and requested a way to
not break Mageia.

After discussing with Florian Festi about it, Mageia's
pythonegg(X)(M) will be supported by adding '--legacy'
as a switch to generate legacy Provides to maintain
compatibility with Mageia's existing usage and to give
them a path to transition from this usage over time.

This switch will also enable pythonXdist(M) format
to allow for a transition, as the conditions for
using pythonXdist(M) are the same as pythonegg(X)(M).