Put all Perl-related macros in a single file.
If the '%perl_...' macros are '%define'd in 'macros[.in]', they are
automatically loaded on startup, so '%include'ing the extra
'macros.perl' in the spec file is redundant.

In case that the '%perl_...' macros are sufficiently important to be
'%define'd automatically on startup, the separate file 'macros.perl' can
savely be removed from the distribution.