te: Don't reload header for JUSTDB transactions
I'm working on some custom rpm-ostree code to "install" rpms by
unpacking them with libarchive (since this can be a lot safer than
invoking librpm - my process has total control over writes), then just
using librpm to update the database via _JUSTDB.

However, it appears that for some reason that if the transaction is
not _TEST, we attempt to reload the header from the original package
here.  I don't have the original package, just the header, since they
get unpacked into OSTree commits (with the header as ostree commit
metadata).

Now, I don't understand why we re-parse the RPM - what's wrong with
the original header we loaded?, but it fixes the problem for me to
also skip with _JUSTDB, and should be a very low risk change.