Working '%__scm_apply_quilt' macro.
This nacro is called in '%apply_patch' as the recipient of a pipe.
Unfortunately, 'quilt import' only works on 'real' files and does not
read from STDIN (like 'quilt import -'). We circumvene this restriction
with a temporary file(name).

This modified macro has successfully been tested on Kubuntu 12.04.5 with
'quilt 0.50' and 'debbuild 16.1.6' (https://github.com/ascherer/debbuild).