ndb: add rpmpkgLockInternal()
Previously it wasn't able to get an exclusive lock with rpmpkgLock() if
pkgdb was open readonly. That's ok for pkgdb  but rpmpkgLock() is also
used in rpmxdbLock() which is used for locking indexes. And when pkgdb
was open rdonly even if indexes were open rw, it wasn't possible to
write into them because it wasn't possible to get exclusive lock with
rpmxdbLock()->rpmpkgLock().

Having pkgdb readonly and indexes rw can happen when pkdb is open as
readonly but during opening of indexes some missing index is detected
so it is necessary to open indexes as rw and regenerate them.

Now rpmpkgLock() enables to get exclusive lock without having pkgdb
open as rw. New function rpmpkgLockInternal() behaves as rpmpkgLock()
previously and it is used only for pkgdb.