ndb: fix regression from 20f2c51053c1b2e302eff8240613568443bbe98b
If there was need to regenerate indexes and index file was open readonly
then index file was closed and open again as readwrite. But problem was
that index file may have already been memory mapped so closing it wasn't
right solution.

Now if it is possible then index file is open as readwrite from
the beggining no matter of requested open flags.