Current plugins don't make sense in rpmbuild so disable them there
In rpmbuild only an init hook was called and no other hooks was called.
So the plugins were initialized in rpmbuild but they were never used
in rpmbuild. Even the init hook was called from place where calling
of init hook made no sense (it was called from rpmCheckDeps()).

Also it wasn't possible to disabled plugins in rpmbuild with --noplugins,
which caused problems in rpm test suite.