Fix rpmi.at testcase by checking actual datadir and docdir used.
The testcase for rpm -i with/without --excludedocs had hardcoded paths
checking if files were installed or not. This meant that depending on
the configure --prefix used the test might fail because the datadir
and/or docdir were different from expected. Fixed by checking the
actual datadir and (default)docdir to test against.