Make creating index records consistent for rich and rich-weak deps
If a package contains a rich require dependency then this rich
dependency is parsed and also subdependencies are stored into the
require index.  For example for rich dependency "(A and B)" the whole
string "(A and B)" is stored into the index Requirename but
subdependcies "A" and "B" are stored into index too. Previously this
parsing and storing subdependencies was done only for require rich
dependencies. Now it is done also for weak rich dependecies (suggests,
supplements and recommends).
(rhbz:1325982)