Fix signing with non-ASCII uid keys (rhbz:1243963)
Removed setting LC_ALL to "C" because since commit [1] the gpg program
gets password by yourself from terminal so there is no sense in
setting LC_ALL to "C" if the terminal settings is e. g. UTF-8. That was
only confusing gpg program and it was not able to properly get and
display non-ASCII characters.

[1] 0bce5fcf270711a2e077fba0fb7c5979ea007eb5