rpmrc: Convert uname.machine == "riscv" to "riscv32"/"riscv64"/"riscv128".
On RISC-V, the kernel can return uname.machine == "riscv" (for all bit
sizes).  I say "can" return, because that is the default, but it is
also possible to compile the kernel specially so it returns "riscv64"
etc.

GNU is using "riscv64".

This commit converts the kernel uname machine type "riscv" to a more
suitable value.

This conversion is supposed to be done by the arch_canon table.
However the arch_canon table is not populated until after the
defaultMachine function is called for the first time, making it a bit
useless.  In any case, arch_canon cannot take into account the bit
size of the architecture, but the C code here can.

Signed-off-by: Richard W.M. Jones <rjones@redhat.com>