rpmrc: Convert uname.machine == "riscv" to "riscv32"/"riscv64"/"riscv128".
On RISC-V, the kernel returns uname.machine == "riscv" (for all bit
sizes).  GNU is using "riscv64".  Convert the kernel uname machine
type to a more suitable value.

This conversion is supposed to be done by the arch_canon table.
However the arch_canon table is not populated until after the
defaultMachine function is called for the first time, making it a bit
useless.  In any case, arch_canon cannot take into account the bit
size of the architecture, but the C code here can.