Put %GNUconfigure macro out of its misery
Nobody can be using this thing because, well, its so broken its not
possible to use it (see RhBug:1271275). In fact, it was already deemed
that way 15 years ago in RhBug:56189. The thing hasn't been touched
at all in 16 years so lets just kill it already, lest we have a
third similar bug in another 15 years.