Always sanity check the header structure as the first thing
If there are inconsistencies in the header we don't want to touch
it no matter what it's signed with. Or not. headerVerifyInfo()
is not a very thorough sanity check as it is but its better than not
doing it at all, the height of embarrasment is the cases where
--nosignature --nodigests mode catches a bogus header but trying
to verify signature/digests segfaults...