Don't bother headerVerifyInfo() with region trailer tag
The region trailer tag is so different from everything else it
makes little sense to call headerVerifyInfo() on it. We're validating
all the values locally here anyway. Except for negative region index
length, which headerVerifyInfo() used to catch so now we need to
check manually.

This eliminates the need for the hysterical "negate" argument
to headerVerifyInfo().