Don't look for merged signatures when reading package files
When reading package files, signatures are being checked elsewhere
and exact_size is enabled and il > ril wont be permitted this far.
In particular there must NOT be any dribbles or dribble signatures
present at this point in package file read so dont expose the code
to what isn't even supposed to be there.