Force header data formatting through a central function
Make rpmHeaderFormatFuncByName() and rpmHeaderFormatFuncByValue() return
an opaque pointer to the extension itself (instead of the actual function
pointer) and drop Func from the function and type names to signify the
point. Add rpmHeaderFormatCall() function through which all formatting
has to be done now.

This is not supposed to change anything in itself, but opens up
possibilities for central error checking etc.