Make headerVerifyInfo() return code sane
Return zero for zuccess (haha) like most of the world does, -1
for success is so bizarre it's like violating the principle of
least surprise on purpose. Callers only ever cared about the
non-success value for error reporting but since that is taken
care of by headerVerifyInfo() it doesn't matter at all. Return
the ordinal of failing tag achieves the same should anybody care.

Leaving the error message to actual index number to keep the message
compatible, not that its likely to matter.