Always sanity check the entire header, not just immutable region
To begin with, rpm v3 packages don't HAVE an immutable region. Which
means that for v3 packages whose signatures we cannot verify on normal
paths, we'd also miss even the most basic of sanity checks because
ril on v3 packages is zero.

For installed packages this also enables data overlap checks for
tags outside the region, which up to now have been checked but only
one by one in headerSigVerify().