Fix %ghost verification and --noghost behavior (RhBug:1395818)
Verify has "always" accepted --noghost as an option but it's always
been broken too, the --noghost option used to *clear* the bit on
qva_fflags, bit which nothing set in the first place. When you
test that bit for enabling verify on ghosts, chances are the ghosts
remain ethereal...

That was until commit efd696d32d343b82832dd9e4aee992bff8e64ad0 fixed
the bit clear to set to make --noghost work on queries, but hardwired
the verify code to no ghosts. Wtf?

So... after 15 years of existence (commit
60977b6c27aca83d1c15deef2e4f759268308075) and somewhat colorful
history, lets make it work and document it too. qva_fflags is supposed
to be a "filter out if matching attribute in file" bitfield but its been
kinda hard to tell when there was one example left which had it the wrong
way around. The original commit has two more examples where the bit setting
is right but disabled with "ifdef NEVER" ... since removed.