Postpone all signature tag content validation to rpmSigInfoParse()
The well-intended checks in headerSigVerify() have been a source
of trouble more than once, and for example the SHA1 check was still
assuming \0-terminated string *before* checking the type. The current
checks in rpmSigInfoParse() might not be perfect but consolidating
everything there at least means we dont have to fix the same thing
fifteen times all over the code.