Consolidate signature + header read to new hdrblobRead() function
The signature and package header reading only differ by a few minor
bits like number of tags allowed, doing this in copy-slop code is
just dumb.

There's an ugly little quirk here in that headerImport() can in theory
still fail even if hdrblobRead() succeeds and we need to manually
clean up in that case. But that can be addressed in a later commit...