Handle signature header padding in hdrblobRead()
We already know whether its a signature or not so it only makes sense
to do the rest of the IO there too to avoid annoyances like signature
header read + import succeeding but then failing on invalid padding etc.

rpmReadSignature() does an fstat() and debug printing in addition,
leaving that there at the moment as it requires an imported header.

It's worth noting that for v3 packages headerSizeof() the returned
size (both il + dl) is larger than what il/dl off the disk says because
a legacy region tag 'n stuff is inserted on import. The difference
is exactly 32 bytes so as it happens it doesn't affect our padding,
but its a quirk alright.