Clean up headerWrite() a bit
A switch-case for a boolean flag whether to do something or not is
rather hysterical. A good old "if" is shorter and actually much
easier to read. When writing the magic, we know we're writing exactly
one of them, shuffle the arguments to better match Fwrite() interface.
Initialize uh on declaration line to save an extra line. No actual
functional changes.