When using %autopatch, create backup files with .<num>~ suffix by default
The legacy mandriva %apply_patches macro by default  generated backup
files with unique suffixes by default, the suffix for patch0 being
.0001~ etc.

As for those of us still using gendiff, this is still convenient, while
for those using a scm, the '~' at end of suffix will anyways be
automatically ignored, so enabling it by default won't interfer.

A -B argument passed to the %autopatch will disable this.