Do not expand macro in a comment.
Although %% can disable macro expansion, requiring to add another %
to % in a commented-out lines is not intuitive to most developers and
we are observing many mistakes in spec files.

With multi-line macros, expanding macros in a comment is disasterous
and is hard to find the errors for developers who do not know that
using % in # may be disasterous, which is very not common for other
scripting languages.

Because expanding a single-line macro in a comment has no effect on
the behaviors of rpm-build and expanding multi-line macro in a comment
has no value as a scripting syntax, I'd like to suggest that we do not
expand macros in a comment.

Signed-off-by: MyungJoo Ham <myungjoo.ham@samsung.com>