Error if a macro in a comment is expanded to multiple lines
Multi-line macro expansion in a comment has side effects
as mentioned in the PR #123. However, because there are
cases where a macro is needed to be expanded at a line
starting with # (e.g., shebang), let's emit an error
in a case where it is expected to cause side effects while
we do not expect to have any useful effects.

Related issue: #121

This is an alternative to #123 according to the comment from
@pmatilai in #123.

Signed-off-by: MyungJoo Ham <myungjoo.ham@samsung.com>