[WIP] Replace gzip/bzip2/xz with pigz/pbzip2/pxz for multicore environment
This PR is to fix issue #113. And It can be used as an alternative of PR #117.
In modern generation, most of the developers are using a machine based on
multi-core architecture. From now on, Let's prepare to support parallelism.

pigz is compatible with the existing gzip because pigz is using gzip's library.
$ ldd /usr/bin/pigz | grep -e 'libz\|pthread'
	libpthread.so.0 => /lib/x86_64-linux-gnu/libpthread.so.0 (0x00007fe318f23000)
	libz.so.1 => /lib/x86_64-linux-gnu/libz.so.1 (0x00007fe318d09000)

pbzip2 is compatible with the existing pbzip2 because pbzip2 is using bzip2's library.
$ ldd /usr/bin/pbzip2 | grep -e 'libbz2\|pthread'
	libbz2.so.1.0 => /lib/x86_64-linux-gnu/libbz2.so.1.0 (0x00007fa9d089a000)
	libpthread.so.0 => /lib/x86_64-linux-gnu/libpthread.so.0 (0x00007fa9d067d000)

pxz is compatible with the existing xz because pxz is using xz's library.
$ ldd /usr/bin/pxz |grep -e 'liblzma\|pthread'
	liblzma.so.5 => /lib/x86_64-linux-gnu/liblzma.so.5 (0x00007f27d3fd5000)
	libpthread.so.0 => /lib/x86_64-linux-gnu/libpthread.so.0 (0x00007f27d3b95000)

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>