unconditionally build embedded python interpreter with bindings
Considering that the actual code for it is now compiled into the rpm
python extension rather than an additional separate library and that the
code size for is neglible, it's less messy with no good arguments
against why making it unconditional.

Additionally it avoids extra hackery needed for dealing with config
definitions outside of the python module root when module is built with
distutils (setup.py).