Fix thinko vrt unknown and other signature types in rpmpkgVerifySigs()
"Other" signature types are sizes and the like whose values we don't
care about in here, but whose types we validate and want to be sane.
So take errors from rpmSigInfoParse() into account, but skip
rpmVerifySignature() because it doesn't make sense for the "other" types.

On the other hand, unknown types we should just skip, this is necessary
to avoid unnecessarily breaking forward-compatibility.